@@ -1,9 +1,7 @@
 import React from 'react';
-import { Link } from 'react-router';
 import classnames from 'classnames/bind';
 
 import Head from '../../components/head/';
-import Titlebar from '../../components/titlebar/';
 import Window from '../../components/window/';
 import Directory from '../../components/directory/';
 
@@ -21,8 +19,8 @@ class Folder extends React.Component {
     return (
       <section className={className}>
         <Head data={data} />
-        <Window name={data.name} parent={data.parent}>
-          <Directory data={data.items} />
+        <Window data={data}>
+          <Directory data={data.items} cssClasses="c-directory--wrap" />
           {this.props.children}
         </Window>
       </section>
@@ -33,16 +31,7 @@ class Folder extends React.Component {
 Folder.propTypes = {
   route: React.PropTypes.shape({
     data: React.PropTypes.shape({
-      name: React.PropTypes.string,
-      description: React.PropTypes.string,
-      url: React.PropTypes.string,
-      image: React.PropTypes.string,
-      items: React.PropTypes.arrayOf(React.PropTypes.shape({
-        id: React.PropTypes.number,
-        name: React.PropTypes.string,
-        url: React.PropTypes.string,
-        type: React.PropTypes.string,
-      })),
+      parent: React.PropTypes.string,
     }),
   }),
   children: React.PropTypes.element,
