@@ -8,16 +8,14 @@ import styles from './style.css';
 
 const cx = classnames.bind(styles);
 
-const Window = ({ parent, name, children, cssClasses }) => {
+const Window = ({ data, children, cssClasses }) => {
   const className = cx('c-window', cssClasses);
 
   return (
     <div className={className}>
       <div className={`${cx('c-window__header')}`}>
-        <Link to={parent}>
-          <Icon name="close" />
-        </Link>
-        {name}
+        <Link to={data.parent}><Icon name="close" /></Link>
+        {data.name} {data.memory.full} full, {data.memory.free} free, {data.memory.use} in use
       </div>
       <div className={`${cx('c-window__main')}`}>
         {children}
@@ -27,10 +25,26 @@ const Window = ({ parent, name, children, cssClasses }) => {
 };
 
 Window.propTypes = {
-  parent: React.PropTypes.string,
-  name: React.PropTypes.string,
+  data: React.PropTypes.shape({
+    name: React.PropTypes.string,
+    parent: React.PropTypes.string,
+    memory: React.PropTypes.shape({
+    }),
+  }),
   children: React.PropTypes.node,
   cssClasses: React.PropTypes.string,
+};
+
+Window.defaultTypes = {
+  data: {
+    name: '',
+    parent: '',
+    memory: {
+      full: '',
+    },
+  },
+  children: '',
+  cssClasses: '',
 };
 
 export default Window;
