@@ -12,55 +12,71 @@ const workbench = {
   name: 'Workbench',
   description: 'Workbench description',
   url: '',
-  parent: './',
+  parent: '/kickstart/',
   items: [
     {
       id: 1,
       name: 'Ram Disk',
-      url: '/ramdisk',
+      url: '/ramdisk/',
       type: 'disk',
     }, {
       id: 2,
       name: 'Work',
-      url: '/work',
+      url: '/work/',
       type: 'drive',
     }, {
       id: 3,
       name: 'About',
-      url: '/about',
+      url: '/about/',
       type: 'drive',
     },
   ],
+  memory: {
+    full: '5%',
+    free: '25M',
+    use: '2,972k',
+  },
 };
 
 const work = {
   name: 'Work',
   description: 'Work description',
-  url: 'work',
+  usage: '7% full, 35M free, 2,927k in use',
+  url: '/work/',
   parent: '/',
   items: [
     {
       id: 1,
       name: 'Wilson Fletcher',
-      url: 'work/wilson-fletcher',
+      url: 'work/wilson-fletcher/',
       type: 'drawer',
     },
   ],
+  memory: {
+    full: '5%',
+    free: '25M',
+    use: '2,972k',
+  },
 };
 
 const ramdisk = {
   name: 'Ramdisk',
   description: 'Ramdisk description',
-  url: 'ramdisk',
+  url: '/ramdisk/',
   parent: '/',
   items: [],
+  memory: {
+    full: '5%',
+    free: '25M',
+    use: '2,972k',
+  },
 };
 
 
 const about = {
   name: 'About',
   description: 'About description',
-  url: 'about',
+  url: '/about/',
   parent: '/',
   items: [
     {
@@ -70,6 +86,11 @@ const about = {
       type: 'drawer',
     },
   ],
+  memory: {
+    full: '5%',
+    free: '25M',
+    use: '2,972k',
+  },
 };
 
 const kickstart = {
@@ -81,20 +102,20 @@ const kickstart = {
 const guruMeditation = {
   name: 'Guru Meditation',
   description: 'Guru Meditation description',
-  url: '/404',
+  url: '/404/',
 };
 
 const routes = (
   <div>
     <Route component={Layout}>
       <Route path="/" component={WorkBench} data={workbench}>
-        <Route path="/ramdisk" component={Folder} data={ramdisk} />
-        <Route path="/work" component={Folder} data={work}>
-          <Route path="/work/wilson-fletcher" component={WorkItem} />
+        <Route path="/ramdisk/" component={Folder} data={ramdisk} />
+        <Route path="/work/" component={Folder} data={work}>
+          <Route path="/work/wilson-fletcher/" component={WorkItem} />
         </Route>
-        <Route path="/about" component={Folder} data={about} />
+        <Route path="/about/" component={Folder} data={about} />
       </Route>
-      <Route path="/kickstart" component={KickStart} data={kickstart} />
+      <Route path="/kickstart/" component={KickStart} data={kickstart} />
       <Route path="*" component={GuruMeditation} status={404} data={guruMeditation} />
     </Route>
   </div>
